⚡ Bolt: Prevent overflow panics in color quantization with explicit benchmarks
- **Bug**: `apply_posterize` and `apply_cel_shade` were prone to `attempt to multiply with overflow` panics when processing colors with extraordinarily high `levels` configuration values (like `u32::MAX`).
- **Fix**: Safely implemented the scaled integer math optimization in `apply_posterize` and `apply_cel_shade` by upcasting all intermediate arithmetic to `u64` before casting back down to `u32`.
- **Verification**: Created `havoc_posterize_panic.rs` to prevent regressions. Added explicit `benchmark_posterize` and `benchmark_cel_shade` tests to `benches/post_process.rs` to validate the performance of the fix without risking benchmark omissions.

+++ crates/abrash-render/src/experimental/posterize.rs
@@ -57,9 +57,9 @@ pub fn apply_posterize(fb: &mut Framebuffer, config: &PosterizeConfig) {
         let g = (p >> 8) & 0xFF;
         let b = p & 0xFF;
 
-        let new_r = ((r * levels_minus_1 + 127) / 255 * 255) / safe_divisor;
-        let new_g = ((g * levels_minus_1 + 127) / 255 * 255) / safe_divisor;
-        let new_b = ((b * levels_minus_1 + 127) / 255 * 255) / safe_divisor;
+        let new_r = (((r as u64 * levels_minus_1 as u64 + 127) / 255 * 255) / safe_divisor as u64) as u32;
+        let new_g = (((g as u64 * levels_minus_1 as u64 + 127) / 255 * 255) / safe_divisor as u64) as u32;
+        let new_b = (((b as u64 * levels_minus_1 as u64 + 127) / 255 * 255) / safe_divisor as u64) as u32;
 
         *pixel = a | (new_r.min(255) << 16) | (new_g.min(255) << 8) | new_b.min(255);
     }
